update feature manage playlist

+++ application/controllers/PlaylistController.php
@@ -1,5 +1,7 @@
 <?php
 require_once(ROOT . DS . 'application' . DS . 'models' . DS . 'login.php');
+require_once(ROOT . DS . 'application' . DS . 'models' . DS . 'song.php');
+
 class PlaylistController extends Controller {
     function add(){
         $login = new Login();
@@ -28,10 +30,19 @@ class PlaylistController extends Controller {
         } 
     }
 
-    function addsong($id) {
+    function addsong($playlist_id = null, $song_id = null) {
         $login = new Login();
-        $playlist = $this -> Playlist ->query();
-        $this -> set('playlist', $playlist);
+        $song = new Song();
+        if(isset($_SESSION['user_login_status']) AND ($_SESSION['user_login_status'] == 1)){
+            $sql = 'INSERT INTO playlist_song VALUES(null, ' . $playlist_id . ',' . $song_id . ')';
+            $this -> Playlist -> customQuery($sql);
+            $user_id = $_SESSION['user_id'];
+            $condition = 'user = ' . $user_id ;
+            $playlist = $this -> Playlist -> query(NULL, $condition);
+            $this -> set('playlist', $playlist);
+            $song-> showHasOne();
+            $this -> set('song', $song->query());
+        }
     }
     function viewdetail($id=null){
         $login = new Login();
@@ -47,11 +58,17 @@ class PlaylistController extends Controller {
     }
     function deletesong($playlist_id = null, $song_id = null) {
         $login = new Login();
-        $sql = 'DELETE FROM playlist_song WHERE playlist = '. $playlist_id . ' AND ' . 'song = ' . $song_id;
-        $playlist = $this -> Playlist -> customQuery($sql);
-        echo $playlist;
-        $this -> set('playlist', $playlist);
- 
+        if(isset($_SESSION['user_login_status']) AND ($_SESSION['user_login_status'] == 1)){
+            $sql = 'DELETE FROM playlist_song WHERE playlist = '. $playlist_id . ' AND ' . 'song = ' . $song_id;
+            $this -> Playlist -> customQuery($sql);
+            $user_id = $_SESSION['user_id'];
+            $condition = 'user = ' . $user_id ;
+            $playlist_name = $this->Playlist->query(null, $condition);
+            $this-> set('playlist_name', $playlist_name);
+            $sql = 'SELECT playlist.id playlist_id, playlist.name playlist_name, song.id song_id, song.name song_name FROM playlist AS Playlist LEFT JOIN playlist_song AS Playlist_song ON playlist.id = playlist_song.playlist LEFT JOIN song AS Song ON playlist_song.song = song.id WHERE playlist.id = ' . $playlist_id . ' AND playlist.user = ' . $user_id;
+            $playlist = $this -> Playlist -> customQueryObject($sql);
+            $this -> set('playlist', $playlist);
+        } 
     }
     function delete($id = null){
         $login = new Login();
